Move font related code to separate location
To avoid already huge agg_image.cpp file growing even further.
